[Silver I] Title: 삼각 그래프, Time: 32 ms, Memory: 4364 KB -BaekjoonHub

+++ 백준/Silver/4883. 삼각 그래프/삼각 그래프.cc
@@ -7,8 +7,8 @@
 using namespace std;
 
 int N;
-long long board[100001][3];
-long long dp[100001][3];
+int board[100001][3];
+int dp[100001][3];
 const int INF = 0X7FFFFFFF;
 
 void Input()
